Code changes to make it (almost) compile on MacOS X version 10.2.3 with GCC version "gcc (GCC) 3.1 20020420 (prerelease)". The changes are:
 * Most of the changes are to fix the arguments to printf(3)
   and friends when printing size_t argument:
   OLD:  printf("%d\n", my_list.size());
   NEW:  printf("%u\n", (uint32_t)my_list.size());

 * Explicitly cast timeval.tv_sec and timeval.tv_usec
   to (unsigned long) when used inside printf(3) and friends.

 * Explicit cast when assigning to msghdr.msg_name:
   OLD: mh.msg_name = &foo;
   NEW: mh.msg_name = (caddr_t)&foo;

   On FreeBSD and Linux, msghdr.msg_name is of type (void *), but
   on OS X it is of type (char *), therefore the need for explicit
   cast.

 * Change "template <class _A>" to "template <class A>",
   and propagate the change inside the template,
   because the compiler didn't like underscore in front of
   the class name.

 * De-constify the argument to ether_ntoa(3):
   OLD: ether_ntoa(&ea);
   NEW: ether_ntoa(const_cast<struct ether_addr *>(&ea));

   The reason is because OS X has a broken ether_ntoa() declaration
   in <net/ethernet.h>, even though the manual page shows
   the argument is const.

 * Use "#ifdef SIOCGETSGCNT_IN6" to isolate code that uses
   ioctl(SIOCGETSGCNT_IN6). This is IPv6 multicast-related,
   and even though it exists in KAME, somehow it is missing
   from OSX, even though it appears that OSX is using KAME's
   IPv6 code. Currently, if ioctl(SIOCGETSGCNT_IN6) is missing,
   the multicast code will abort at run time. In the future,
   the solution should be to detect this in compilation time,
   and to completely disable IPv6 compilation of the multicast
   code.

 * Use #ifndef SIOCGETMIFCNT_IN6 to isolate code that
   uses ioctl(#ifndef SIOCGETMIFCNT_IN6).
   See the above comment about SIOCGETSGCNT_IN6 for further
   details.

 * Use #ifdef HOST_OS_MACOSX to isolate code that uses
   "struct in6_ifreq".
   See the above comment about SIOCGETSGCNT_IN6 for further
   details.

The only place where the compilation fails is in
bgp/harness/command.cc in assigments like:
  _commands["reset"] = &Command::reset;
However, I think this is a problem with the compiler, hence
I didn't try to modify the code.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@225 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/bgp/aspath.cc
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/bgp/aspath.cc,v 1.4 2003/01/21 01:31:49 rizzo Exp $"
+#ident "$XORP: xorp/bgp/aspath.cc,v 1.5 2003/01/26 01:22:35 mjh Exp $"
 
 // #define DEBUG_LOGGING
 // #define DEBUG_PRINT_FUNCTION_NAME
@@ -68,7 +68,7 @@ AsSegment::_encode(size_t &len) const
 	len += (iter->is_extended()) ? 4 : 2;
     assert(_aslist.size() == _entries);		// XXX this is expensive
 
-    debug_msg("data size = %d\n", len);
+    debug_msg("data size = %u\n", (uint32_t)len);
     uint8_t *data = new uint8_t[len];
 
     size_t pos = 0;
@@ -214,7 +214,7 @@ AsPath::AsPath(const char *as_path) throw(InvalidString)
     for (size_t i = 0; i < path.length(); i++) {
 	char c = path[i];
 
-        debug_msg("check <%c> at pos %d\n", c, i);
+        debug_msg("check <%c> at pos %u\n", c, (uint32_t)i);
 	if (isdigit(c)) {
 	    size_t start = i;
 
@@ -287,8 +287,8 @@ AsPath::add_segment(const AsSegment& s)
 
     size_t n = s.get_as_path_length();
     _path_len += n;
-    debug_msg("End of add_segment, As Path length by %d to be %d\n",
-                n, _path_len);
+    debug_msg("End of add_segment, As Path length by %u to be %u\n",
+                (uint32_t)n, (uint32_t)_path_len);
 }
 
 string
