+++ trunk/xorp/bgp/notification_packet.cc
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/bgp/notification_packet.cc,v 1.8 2003/01/22 02:46:34 rizzo Exp $"
+#ident "$XORP: xorp/bgp/notification_packet.cc,v 1.9 2003/01/24 19:50:10 rizzo Exp $"
 
 #include "bgp_module.h"
 #include "config.h"
@@ -28,8 +28,8 @@ NotificationPacket::NotificationPacket(uint8_t		ec,
 				       const uint8_t*	ed,
 				       size_t		elen)
 {
-    debug_msg("Error code %d sub code %d data %#x len %d\n", ec, esc,
-	      (int)ed, elen);
+    debug_msg("Error code %d sub code %d data %#x len %u\n", ec, esc,
+	      (int)ed, (uint32_t)elen);
 
     if (elen == 0)
 	ed = 0;
